+++ framework/db-local/src/test/java/is/codion/framework/db/local/TestDomain.java
@@ -194,12 +194,12 @@ public final class TestDomain extends DomainModel {
 														.caption(Employee.SALARY.name())
 														.nullable(false)
 														.range(1000, 10000)
-														.maximumFractionDigits(2),
+														.fractionDigits(2),
 										Employee.COMMISSION.define()
 														.column()
 														.caption(Employee.COMMISSION.name())
 														.range(100, 2000)
-														.maximumFractionDigits(2),
+														.fractionDigits(2),
 										Employee.MGR.define()
 														.column(),
 										Employee.MGR_FK.define()
@@ -293,7 +293,7 @@ public final class TestDomain extends DomainModel {
 										EmployeeFk.SALARY.define()
 														.column()
 														.caption(EmployeeFk.SALARY.name())
-														.maximumFractionDigits(2),
+														.fractionDigits(2),
 										EmployeeFk.COMMISSION.define()
 														.column()
 														.caption(EmployeeFk.COMMISSION.name()),