+++ framework/domain/src/main/java/is/codion/framework/domain/entity/package-info.java
@@ -22,7 +22,7 @@
  * <li>{@link is.codion.framework.domain.entity.EntityValidator#STRICT_VALIDATION}
  * <li>{@link is.codion.framework.domain.entity.Entities#VALIDATE_FOREIGN_KEYS}
  * <li>{@link is.codion.framework.domain.entity.Entities#STRICT_DESERIALIZATION}
- * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#MAXIMUM_FRACTION_DIGITS}<br>
+ * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#FRACTION_DIGITS}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#TIME_FORMAT}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#NUMBER_FORMAT_GROUPING}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#GROUPING_SEPARATOR}<br>
@@ -30,7 +30,7 @@
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#USE_LEXICAL_STRING_COMPARATOR}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#DATE_TIME_FORMAT}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#DATE_FORMAT}<br>
- * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#DECIMAL_ROUNDING_MODE}<br>
+ * <li>{@link is.codion.framework.domain.entity.attribute.AttributeDefinition#ROUNDING_MODE}<br>
  * <li>{@link is.codion.framework.domain.entity.attribute.ForeignKeyDefinition#REFERENCE_DEPTH}
  * </ul>
  */