+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/Chinook.java
@@ -302,7 +302,7 @@ public final class Chinook extends DomainModel {
 				Invoice.TOTAL.define()
 					.column()
 					.caption("Total")
-					.maximumFractionDigits(2))
+					.fractionDigits(2))
 			.keyGenerator(identity())
 			.caption("Invoice")
 			.build();