+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/Petstore.java
@@ -63,12 +63,12 @@ public final class Petstore extends DomainModel {
 					.column()
 					.caption("Latitude")
 					.nullable(false)
-					.maximumFractionDigits(2),
+					.fractionDigits(2),
 				Address.LONGITUDE.define()
 					.column()
 					.caption("Longitude")
 					.nullable(false)
-					.maximumFractionDigits(2),
+					.fractionDigits(2),
 				Address.LOCATION.define()
 					.column()
 					.caption("Location"),
@@ -297,7 +297,7 @@ public final class Petstore extends DomainModel {
 					.column()
 					.caption("Price")
 					.nullable(false)
-					.maximumFractionDigits(2),
+					.fractionDigits(2),
 				Item.ADDRESS_ID.define()
 					.column()
 					.nullable(false),