+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/World.java
@@ -149,7 +149,7 @@ public final class World extends DomainModel {
 					.column()
 					.caption("Surfacearea")
 					.nullable(false)
-					.maximumFractionDigits(2),
+					.fractionDigits(2),
 				Country.INDEPYEAR.define()
 					.column()
 					.caption("Indepyear"),
@@ -160,15 +160,15 @@ public final class World extends DomainModel {
 				Country.LIFEEXPECTANCY.define()
 					.column()
 					.caption("Lifeexpectancy")
-					.maximumFractionDigits(1),
+					.fractionDigits(1),
 				Country.GNP.define()
 					.column()
 					.caption("Gnp")
-					.maximumFractionDigits(2),
+					.fractionDigits(2),
 				Country.GNPOLD.define()
 					.column()
 					.caption("Gnpold")
-					.maximumFractionDigits(2),
+					.fractionDigits(2),
 				Country.LOCALNAME.define()
 					.column()
 					.caption("Localname")
@@ -220,7 +220,7 @@ public final class World extends DomainModel {
 					.column()
 					.caption("Percentage")
 					.nullable(false)
-					.maximumFractionDigits(1))
+					.fractionDigits(1))
 			.caption("Countrylanguage")
 			.build();
 	}